Move paths to separate ix_apps package
