+++ src/middlewared/middlewared/plugins/apps/app_query_utils.py
@@ -2,7 +2,7 @@ import os
 from dataclasses import dataclass
 
 from .app_utils import get_app_metadata
-from .app_path_utils import get_app_parent_config_path
+from .ix_apps.path import get_app_parent_config_path
 from .docker.query import list_resources_by_project
 from .utils import PROJECT_PREFIX
