+++ src/middlewared/middlewared/plugins/apps/app_utils.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 from semantic_version import Version
 
-from .app_path_utils import get_installed_app_metadata_path, get_installed_app_versions_dir_path
+from .ix_apps.path import get_installed_app_metadata_path, get_installed_app_versions_dir_path
 
 
 app_details = namedtuple('app_details', ['name', 'version'])
